### 2025-11-04 #9250 @crvernon
@mutopsy you state that the package has the license GPL-3 but you need a formal license file in the root of your repository that directly matches the language specified from here:  https://opensource.org/licenses
