+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-10-14T01:42Z
+last_built: 2025-11-05T08:14Z
 urls:
   reference: https://mutopsy.github.io/pwranova/reference
   article: https://mutopsy.github.io/pwranova/articles
