### 2026-01-19 cf89d7c
fixed denominator df for paired contrast tests.
