+++ man/pwrcontrast.Rd
@@ -52,7 +52,7 @@ or paired/repeated-measures settings.
 \details{
 For a contrast with weights \eqn{w_1, \dots, w_K} that sum to zero,
 the numerator df is 1. The denominator df is \eqn{n - K} for
-between-subjects (unpaired) designs and \eqn{(n - 1)(K - 1)} for
+between-subjects (unpaired) designs and \eqn{n - 1} for
 paired/repeated-measures designs. Power uses the noncentral \emph{F}-with
 \eqn{\lambda = f^2 \cdot n_{\mathrm{total}}}.
 
