+++ cran-comments.md
@@ -3,15 +3,7 @@
 * win-builder (devel and release)
 
 ## R CMD check results
-0 errors | 0 warnings | 1 note
-
-## Notes
-* "CRAN incoming feasibility" shows:
-  - Maintainer: 'Hiroyuki Muto <mutopsy@omu.ac.jp>'
-  - New submission
-  This is the expected informational note for initial CRAN submissions.
+0 errors | 0 warnings | 0 note
 
 ## Comments
-This is the first stable release of `pwranova`, an R package for power analysis 
-of flexible ANOVA designs, planned contrasts, *t*-tests, and correlation tests.  
-The package has been validated against G*Power for overlapping designs.
+Bug fix release: corrected denominator degrees of freedom for paired/repeated-measures contrast power calculations.
