### 2026-02-25 #9250 @ruarai
> 👋 [@ruarai](https://github.com/ruarai) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html

Sorry I will have to pass on this one as I am not familiar with power analysis and would not be able to assess the functionality.
