### 2026-02-25 #9250 @AntoineSoetewey
> 👋 @AntoineSoetewey - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html

Dear @adithirgis, I'd be happy to review this submission.
