### 2026-02-27 #10123 @editorialbot
[REVIEW]: pwranova: An R package for power analysis of flexible ANOVA designs and related tests
**Submitting author:** @mutopsy (<a href="http://orcid.org/0000-0002-0007-6019">Hiroyuki Muto</a>)
**Repository:** https://github.com/mutopsy/pwranova
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
