### 2026-02-27 #10123 @adithirgis
👋 @mutopsy , @AntoineSoetewey  , and @RobLBaker - This is the review thread for the paper. All of our communications will happen here from now on.

Please read the following guidelines: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html.
This ensures the review guidelines match the scope under which this submission was received (this submission is being reviewed by the JOSS criteria before this scope update was made on [5th January 2026](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future))
