### 2026-02-27 #10123 @adithirgis
👋 @AntoineSoetewey This review issue has been updated to reflect JOSS's revised scope policies.

What changed:

Documentation links in the reviewer checklist now point to the archived pre-2026 documentation at https://joss.readthedocs.io/en/pre-2026-submission/
This ensures the review guidelines match the scope under which this submission was received
Important context:
JOSS recently updated its scope to address challenges posed by generative AI (see our [blog post](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) for details). However, this submission is being reviewed under the previous scope and should be reviewed according to those guidelines.

The archived documentation preserves the review criteria that were in effect when this paper was submitted. Your existing progress on the checklist has been preserved.
