### 2026-02-27 mutopsy/pwranova#3 @AntoineSoetewey
Add a real-world example (including in the paper)
Ref: https://github.com/openjournals/joss-reviews/issues/10123

First, thank you for making this package available and for the clear examples in the README and website.

From a JOSS review perspective, I wondered whether the package could be strengthened a bit on the “example usage” criterion. I can see examples for `pwranova()` and `pwrcontrast()` in the documentation, but I did not find:

- an example in the paper itself, or
- a worked example framed as a real-world analysis use case.

If you think it fits the scope, it might be helpful to add one applied end-to-end example (brief context, inputs, function call, and short interpretation of results), and include a concise version in the paper as well.

I think this could make the package even easier to adopt for new users and align more closely with JOSS expectations on example usage.
