### 2026-02-27 mutopsy/pwranova#5 @AntoineSoetewey
Add clearer community guidelines (contributing, reporting issues, and support)
Ref: https://github.com/openjournals/joss-reviews/issues/10123

I think the project could be improved with clearer, centralized community guidelines. I may have missed something, but I did not find a single place explaining for external users:

- how to contribute to the project,
- how to report bugs/issues, and
- how to seek help/support.

If helpful, possible additions could be:

- a `CONTRIBUTING.md` describing contribution workflow (coding style, tests, pull request expectations),
- issue templates (or a short section) for bug reports and feature requests,
- a brief “Support” section in README.md indicating where users should ask questions.

I think these additions would make onboarding easier for new contributors and improve the overall user experience.
