### 2026-02-27 mutopsy/pwranova#6 @AntoineSoetewey
Improve high-level summary for non-specialist readers
Ref: https://github.com/openjournals/joss-reviews/issues/10123

I think the package summary could be made even more accessible for non-specialist readers. The current text explains the statistical scope well, but it may feel technical early on for readers who are less familiar with ANOVA terminology.

Would you consider adding a short plain-language summary at the start, for example:

- what practical problem the package helps solve,
- who it is for,
- what kind of output decisions it supports (e.g., planning sample size before data collection).

A brief concrete example sentence (real study-planning context) could also help readers quickly understand the purpose before the technical details.
