### 2026-02-27 mutopsy/pwranova#7 @AntoineSoetewey
Minor writing improvement: split one long sentence for readability
Ref: https://github.com/openjournals/joss-reviews/issues/10123

I have one small writing suggestion to improve readability: the following sentence in the Statement of need may be easier to read if split into two sentences.

Current sentence:

> They also generally do not allow direct specification of user-defined contrasts, and while effect sizes can be specified via Cohen's *f* [@cohen1988statistical], they do not directly support partial eta squared, which is more commonly reported in psychological research as a standard effect size index.

Possible revision:

> They also generally do not allow direct specification of user-defined contrasts. While effect sizes can be specified via Cohen’s *f* [@cohen1988statistical], they do not directly support partial eta squared, which is more commonly reported in psychological research as a standard effect size index.
