### 2026-02-27 mutopsy/pwranova#9 @AntoineSoetewey
Minor references suggestion: consider citing additional related R packages and R itself
Ref: https://github.com/openjournals/joss-reviews/issues/10123

I have a small suggestion regarding references: would it make sense to include citations to a few additional related R packages (beyond `pwr`), where relevant to your comparison scope? For example:

- `WebPower`
- `Superpower`
- `simr`
- `pwr2`
- `pwrss`

I believe this would improve completeness and help readers place the package more clearly within the current landscape.

It may also be helpful to include a citation to R itself (R Core Team).
