### 2026-02-27 mutopsy/pwranova#10 @AntoineSoetewey
Strengthen numeric input validation and finite-value checks
Ref: https://github.com/openjournals/joss-reviews/issues/10123

As a small improvement, would you consider adding stricter validation for numeric inputs (for example, explicit finite-value checks and clearer handling of `Inf`, `-Inf`, and `NaN`), in addition to current range/integer checks?

I think this would make error handling more predictable and improve robustness for edge-case inputs.

Relevant files:

- pwranova.r
- pwrcontrast.r
- pwrttest.r
- pwrcortest.r
