### 2026-02-27 mutopsy/pwranova#11 @AntoineSoetewey
Clarify one-sided test behavior when effect sign is dropped
Ref: https://github.com/openjournals/joss-reviews/issues/10123

I noticed that effect-direction inputs (for example, rho and delta) are converted to absolute values internally. That makes sense for many use cases, but for one-sided alternatives readers may expect directional behavior to be explicit.

Would you consider either:

- documenting this behavior more explicitly in the function docs, or
- preserving sign for one-sided hypotheses (with corresponding tests)?

I think either option would reduce ambiguity for users interpreting one-sided results.

Relevant files:

- pwrttest.r
- pwrcortest.r
