+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2026-01-19T10:40Z
+last_built: 2026-02-28T09:22Z
 urls:
   reference: https://mutopsy.github.io/pwranova/reference
   article: https://mutopsy.github.io/pwranova/articles
