### 2026-02-28 18be502
Add Support section to README
