### 2026-02-28 c6ed7ca
Add CONTRIBUTING.md with contribution guidelines
