+++ CONTRIBUTING.md
@@ -1,0 +1,22 @@
+# Contributing
+
+Contributions are welcome.
+
+## Reporting issues
+
+Please report bugs or request features via GitHub Issues.
+
+## Pull requests
+
+Pull requests are welcome. Please:
+
+- follow existing code style
+- include tests when relevant
+- update documentation if needed
+
+If you are considering major design changes, please start a discussion
+via an issue.
+
+## Questions
+
+For questions, please open an issue.
