### 2026-02-28 4d4bbaf
Clarify bug reporting guidance in README and CONTRIBUTING
