+++ README.md
@@ -138,6 +138,8 @@ https://mutopsy.github.io/pwranova/news/
 If you encounter a bug or would like to request a feature, please open an issue on GitHub:
 https://github.com/mutopsy/pwranova/issues
 
+When reporting a bug, please include a minimal reproducible example if possible.
+
 For questions about usage, feel free to open an issue as well.
 
 ## License
