### 2026-02-28 c10a486
Add example use case section to paper
