### 2026-02-28 3811ecc
Specify required total sample size in example
