+++ paper/paper.md
@@ -65,7 +65,7 @@ pwranova(
 )
 ```
 
-The output provides the minimum sample size required to achieve the desired statistical power, facilitating transparent and well-justified study design. If the `target` argument is omitted, the function returns results for all main effects and interactions under the specified design. More detailed examples and tutorials are available on the package website: 
+The output indicates that a total sample size of 156 is required to achieve the desired statistical power (assuming equal group sizes). If the `target` argument is omitted, the function returns results for all main effects and interactions under the specified design. More detailed examples and tutorials are available on the package website: 
 <https://mutopsy.github.io/pwranova/>
 
 # Acknowledgements
