### 2026-02-28 bfb6f18
Clarify wording for noncentral F-distribution power calculations
