### 2026-02-28 8d4d6d8
Improve sentence flow and coherence
