### 2026-02-28 8d6d46e
Clarify that eta^2_p can be used in place of Cohen’s f in examples
