+++ paper/paper.md
@@ -67,7 +67,7 @@ pwranova(
 
 The output indicates that a total sample size of 156 is required to achieve the desired statistical power (assuming equal group sizes). 
 If the `target` argument is omitted, the function returns results for all main effects and interactions under the specified design. 
-Although Cohen's *f* is used here, $\eta^2_p$ (`peta2`) can also be specified.a
+Although Cohen's *f* is used here, $\eta^2_p$ (`peta2`) can also be specified.a 
 More detailed examples and tutorials are available on the package website: 
 <https://mutopsy.github.io/pwranova/>
 
