### 2026-02-28 mutopsy/pwranova#5 @AntoineSoetewey
Completed via https://github.com/mutopsy/pwranova/commit/c6ed7ca1f293801b63f9f709b08a0e2384bc9463 and https://github.com/mutopsy/pwranova/commit/18be502ed368c965d2632b11c4a1ee2be8c6cff9
