### 2026-03-04 1fe3d87
Add relevant citations
