### 2026-03-04 560d76b
Add relevant citations
