+++ paper/paper.bib
@@ -17,6 +17,53 @@
   url          = {https://CRAN.R-project.org/package=pwr}
 }
 
+@Manual{lu2017pwr2,
+  title        = {pwr2: Power and Sample Size Analysis for One-way and Two-way ANOVA Models},
+  author       = {Pengcheng Lu and Junhao Liu and Devin Koestler},
+  year         = {2017},
+  note         = {R package version 1.0},
+  url          = {https://CRAN.R-project.org/package=pwr2}
+}
+
+@Manual{bulus2025pwrss,
+  title        = {pwrss: Statistical Power and Sample Size Calculation Tools},
+  author       = {Metin Bulus and Sebastian Jentschke},
+  year         = {2025},
+  note         = {R package version 1.0.0},
+  url          = {https://CRAN.R-project.org/package=pwrss}
+}
+
+@book{zhang2018webpower,
+  title   = {Practical Statistical Power Analysis Using Webpower and R},
+  author  = {Zhiyong Zhang and Ke-Hai Yuan},
+  year    = {2018},
+  doi     = {10.35566/power},
+  publisher = {ISDSA Press},
+  address   = {Granger, IN}
+}
+
+@Article{green2016simr,
+  title   = {SIMR: an R package for power analysis of generalized linear mixed models by simulation},
+  author  = {Green, Peter and MacLeod, Catriona J.},
+  journal = {Methods in Ecology and Evolution},
+  year    = {2016},
+  volume  = {7},
+  number  = {4},
+  pages   = {493--498},
+  doi     = {10.1111/2041-210X.12504}
+}
+
+@Article{lakens2021superpower,
+  title = {Simulation-Based Power Analysis for Factorial Analysis of Variance Designs},
+  author = {Dani{\"e}l Lakens and Aaron R. Caldwell},
+  journal = {Advances in Methods and Practices in Psychological Science},
+  year = {2021},
+  volume = {4},
+  number = {1},
+  pages = {2515245920951503},
+  doi = {10.1177/2515245920951503}
+  }
+
 @book{cohen1988statistical,
   title     = {Statistical Power Analysis for the Behavioral Sciences},
   author    = {Cohen, Jacob},
@@ -24,4 +71,13 @@
   edition   = {2nd},
   publisher = {Lawrence Erlbaum Associates},
   address   = {Hillsdale, NJ}
+}
+
+@Manual{R,
+  title        = {R: A Language and Environment for Statistical Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = {2025},
+  url          = {https://www.R-project.org/}
 }
