### 2026-03-04 319cec3
Add a brief example of `pwrcontrast()`
