### 2026-03-04 mutopsy/pwranova#9 @mutopsy
I have added citations to related R packages (WebPower, Superpower, simr, pwr2, and pwrss) in the Statement of Need to better situate the package within the existing ecosystem. I have also added a citation to R (R Core Team). Thanks.
