### 2026-03-05 cd75722
Added `ncp_scale` argument to `pwrcortest()`
