+++ NEWS.md
@@ -1,3 +1,8 @@
+# pwranova 1.1.0 (2026-03-05)
+
+* Added `ncp_scale` argument to `pwrcortest()` for alternative scaling of
+  the noncentrality parameter in correlation power analysis.
+
 # pwranova 1.0.3 (2026-01-19)
 
 * Fixed a bug in the denominator df for paired/repeated-measures contrast tests.
