### 2026-03-05 54c29e0
Strengthen input validation and boundary handling for effect-size conversions and power functions
