### 2026-03-05 faa17d6
Clarify documentation on handling of signed effect-size inputs in pwrcortest() and pwrttest()
