### 2026-03-05 1526c3d
Update README.md
Update README with API overview and G*Power compatibility note
