+++ news/index.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## pwranova 1.1.1 (2026-03-05)
+
+- Strengthened input validation across functions. Added checks for
+  finite values, parameter bounds, and boundary cases in effect-size
+  conversion utilities and power analysis functions.
+- Clarified documentation for
+  [`pwrcortest()`](https://mutopsy.github.io/pwranova/reference/pwrcortest.md)
+  and
+  [`pwrttest()`](https://mutopsy.github.io/pwranova/reference/pwrttest.md)
+  to state that signed effect-size inputs (`rho`, `delta`) are converted
+  to their absolute values.
+
 ## pwranova 1.1.0 (2026-03-05)
 
 - Added `ncp_scale` argument to
