### 2026-03-05 mutopsy/pwranova#11 @AntoineSoetewey
Thanks for the clarification!

Completed via https://github.com/mutopsy/pwranova/commit/faa17d6dab86d6de053ed5a872eeb1a4d7716cce.
