### 2026-03-05 mutopsy/pwranova#10 @AntoineSoetewey
Great, thank!

Completed via https://github.com/mutopsy/pwranova/commit/54c29e0e83c13b18f561b5aac006b0139215f9f8.
