### 2026-05-27 mutopsy/pwranova#21 @smwu
`Documentation`: add note on descriptions for additional arguments
https://github.com/openjournals/joss-reviews/issues/10123

I found the GitHub README "Key arguments" section to be very clear and helpful. One small suggestion: you may want to briefly note that the functions have additional arguments (e.g., `epsilon`, `target`) whose descriptions can be found in the function documentation Help pages. Or you could also consider adding them to the nice argument tables.
