### 2026-06-18 230bdfb
Add GitHub Actions R CMD check workflow
