### 2026-06-18 f79d2d5
Improve README documentation
