+++ index.md
@@ -167,6 +167,10 @@ and the effect-size arguments may also be specified as vectors whose
 length matches the number of ANOVA terms (e.g., main effects or
 interactions).
 
+Additional arguments such as `epsilon` and `target` are described in the
+[`pwranova()` help
+page](https://mutopsy.github.io/pwranova/reference/pwranova.html).
+
 ### pwrcontrast()
 
 The key arguments of
