### 2026-06-18 1138ecf
Clarify interpretation of three-way interaction in example
