+++ paper/paper.md
@@ -61,7 +61,7 @@ especially useful for experimental psychologists and cognitive scientists, as we
 
 # Example use case
 
-To illustrate a realistic application, consider a visual search experiment investigating age-related differences in search efficiency. In this type of task, participants search for a target item among distractors on a screen. They indicate whether the target is present or absent by pressing a key, and response time is recorded. The study includes one between-participant factor, age group (young vs. older adults), and two within-participant factors: target presence (present vs. absent) and set size (8, 16, or 24 items displayed), forming a 2 $\times$ 2 $\times$ 3 mixed design. Suppose the researcher wishes to detect a three-way interaction effect size of *f* = 0.25 with 80% power at $\alpha = .05$. The required total sample size can be estimated as follows:
+To illustrate a realistic application, consider a visual search experiment investigating age-related differences in search efficiency. In this type of task, participants search for a target item among distractors on a screen. They indicate whether the target is present or absent by pressing a key, and response time is recorded. The study includes one between-participant factor, age group (young vs. older adults), and two within-participant factors: target presence (present vs. absent) and set size (8, 16, or 24 items displayed), forming a 2 $\times$ 2 $\times$ 3 mixed design. Suppose the researcher wishes to detect a three-way interaction effect size of *f* = 0.25 with 80% power at $\alpha = .05$. In this example, the three-way interaction reflects whether age-related differences in search efficiency depend on target presence. The required total sample size can be estimated as follows:
 
 ```r
 pwranova(
