### 2026-06-18 054b7b5
Clarify interpretation of three-way interaction in example
