### 2026-06-18 f468c81
Clarify balanced-design assumption in documentation
