### 2026-06-18 5512c7f
Clarify package scope and comparison with existing tools
