+++ paper/paper.bib
@@ -62,7 +62,18 @@
   number = {1},
   pages = {2515245920951503},
   doi = {10.1177/2515245920951503}
-  }
+}
+
+@Article{baranger2023interactionpower,
+  title = {Tutorial: Power analyses for interaction effects in cross-sectional regressions},
+  author = {Baranger, David A. A. and Finsaas, Megan C. and Goldstein, Brandon L. and Vize, Colin E. and Lynam, Donald R. and Olino, Thomas M.},
+  journal = {Advances in Methods and Practices in Psychological Science},
+  year = {2023},
+  volume = {6},
+  number = {3},
+  pages = {1--13},
+  doi = {10.1177/25152459231187531},
+}
 
 @book{cohen1988statistical,
   title     = {Statistical Power Analysis for the Behavioral Sciences},
