### 2026-06-18 be1453d
Add Example use cases in README
