### 2026-06-18 adb6bec
Added specification of the underlying power calculations for ANOVA and t tests.
