### 2026-06-18 3eb8648
Clarified effect-size definitions and noncentrality parameter calculations in pwrttest() documentation.
