### 2026-06-18 403dce2
Clarified effect-size definitions and noncentrality parameter calculations in pwrttest() documentation.
