### 2026-06-18 704932a
Added conventional effect-size benchmarks  in documentation.
