+++ R/pwranova.r
@@ -21,6 +21,9 @@
 #'   as unknown and solved for given \code{n_total}, \code{alpha}, and \code{power}.
 #' @param peta2 Numeric in \eqn{(0,1)}. Partial eta squared. If \code{NULL}, it is
 #'   derived from \code{cohensf} when available.
+#'   For one-way between-subjects ANOVA, \eqn{\eta^2} and \eqn{\eta_p^2}
+#'   are identical. In designs involving multiple effects or repeated measures,
+#'   however, the two measures generally differ.
 #' @param epsilon Numeric in \eqn{(0,1]}. Nonsphericity
 #'   parameter applied to within-subjects terms with \eqn{\mathrm{df}_1 \ge 2}.
 #'   Ignored if there is no within-subjects factor or if all within factors have
@@ -54,6 +57,11 @@
 #'         by GPower (between-, within-, and mixed-factor ANOVA with
 #'         equal cell sizes), \code{pwranova()} was validated to produce results
 #'         identical to those of GPower.
+#'   \item For one-way between-subjects ANOVA, Cohen (1988) suggested
+#'         rough benchmarks of 0.10 (small), 0.25 (medium), and 0.40 (large)
+#'         for \eqn{f}. The corresponding values for \eqn{\eta^2}
+#'         are approximately 0.01, 0.06, and 0.14. These values should
+#'         be regarded as rough guidelines rather than strict criteria.
 #' }
 #'
 #' @return A data frame with S3 class:
