### 2026-06-19 ad58303
Fixed a bug in the noncentrality parameter calculations for within-subjects factors
