### 2026-06-19 3101448
Expanded validation tests, including additional comparisons with G*Power.
