### 2026-06-19 36ff677
Improve documentation of effect sizes and noncentrality parameters
