### 2026-07-07 6372611
Prepare CRAN submission for v1.1.4
