### 2026-07-07 cbb9d21
Added Japanese documentation
