+++ vignettes/README.ja.Rmd
@@ -1,0 +1,222 @@
+---
+title: "pwranovaパッケージの説明 (日本語版)"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{pwranovaパッケージの説明 (日本語版)}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+# pwranova: さまざまなデザインの分散分析や関連する検定の検定力分析を行うためのRパッケージ
+
+`pwranova`は，さまざまなデザインの分散分析 (ANOVA) の検定力分析を行うためのRパッケージです。
+参加者間計画・参加者内計画・混合計画のいずれにも対応しており，**任意の要因数・水準数における主効果と交互作用**の両方に対して検定力分析を実行することができます。
+
+本パッケージでは，ANOVA の各要因および計画対比に対し，検定力・必要な総サンプルサイズ・有意水準・検出可能な最小効果量を計算できます。効果量は偏イータ二乗 ($\eta_p^2$) または Cohen's *f* で指定できます。
+さらに，*t*検定やピアソン相関係数に対する無相関検定など，実験心理学をはじめとする関連分野で広く用いられる検定に対応した検定力分析にも対応しており，包括的な検定力分析パッケージとして利用することができます。
+
+G\*Power (Faul et al., 2009) が対応しているデザインに関しては，本パッケージの計算結果と一致することを確認しています。また，本パッケージは，G\*Powerでは扱えないより柔軟なデザインやオプションにも対応しています。
+本パッケージのすべての分析は，各セルのサンプルサイズが等しい均衡デザイン (balanced design) を前提としています (不均衡デザインに対する検定力分析では一般に，群ごとのサンプルサイズや分散構造について追加の仮定が必要となり，計算が複雑になるからです)。
+
+
+## リンク
+
+- CRAN: <https://CRAN.R-project.org/package=pwranova>
+- ドキュメント (英語版): <https://mutopsy.github.io/pwranova/>  
+- ソースコード: <https://github.com/mutopsy/pwranova/>
+
+## インストール
+
+安定版の`pwranova` は [CRAN](https://CRAN.R-project.org/package=pwranova) に登録されているため，以下のRコードを実行することでインストールできます。
+
+```r
+install.packages("pwranova")
+```
+
+GitHub上の開発版は以下のRコードでインストールできます。 
+
+```r
+# Install devtools if not already installed
+if (!requireNamespace("devtools", quietly = TRUE)) {
+  install.packages("devtools")
+}
+
+# Install pwranova
+devtools::install_github("mutopsy/pwranova")
+```
+
+## 動作環境
+
+- R (>= 4.1.0)
+
+本パッケージは外部パッケージに依存していません。
+
+## クイックスタート
+
+```r
+library(pwranova)
+
+### 1) 検定力の計算 (総サンプルサイズ・効果量・有意水準を指定)
+
+# 参加者間要因が1つ (3水準) ，参加者内要因なし
+res_power_between <- pwranova(
+  nlevels_b = 3,
+  n_total   = 60,
+  cohensf   = 0.25,
+  alpha     = 0.05
+)
+res_power_between # 検定力 = 0.374
+
+# 参加者内要因が1つ (4水準) ，参加者間要因なし
+# (球状性の補正のためにepsilonを指定できます。既定値は1。)
+res_power_within <- pwranova(
+  nlevels_w = 4,
+  n_total   = 30,
[…]
