+++ vignettes/README.ja.Rmd
@@ -222,4 +222,4 @@ GPL-3
 
 ## 引用文献
 
-Faul, F., Erdfelder, E., Buchner, A., & Lang, A.-G. (2009). Statistical power analyses using G\*Power 3.1: Tests for correlation and regression analyses. Behavior Research Methods, *41*(4), 1149–1160. https://doi.org/10.3758/BRM.41.4.1149
+Faul, F., Erdfelder, E., Buchner, A., & Lang, A.-G. (2009). Statistical power analyses using G\*Power 3.1: Tests for correlation and regression analyses. *Behavior Research Methods*, *41*(4), 1149–1160. https://doi.org/10.3758/BRM.41.4.1149
