### 2026-07-11 a852c9c
add doi to some references
