+++ paper/paper.bib
@@ -10,27 +10,30 @@
 }
 
 @Manual{champely2020pwr,
-  title        = {pwr: Basic Functions for Power Analysis},
+  title        = {{pwr}: Basic Functions for Power Analysis},
   author       = {Stephane Champely},
   year         = {2020},
   note         = {R package version 1.3-0},
-  url          = {https://CRAN.R-project.org/package=pwr}
+  url          = {https://CRAN.R-project.org/package=pwr},
+  doi          = {10.32614/CRAN.package.pwr}
 }
 
 @Manual{lu2017pwr2,
-  title        = {pwr2: Power and Sample Size Analysis for One-way and Two-way ANOVA Models},
+  title        = {{pwr2}: Power and Sample Size Analysis for One-way and Two-way ANOVA Models},
   author       = {Pengcheng Lu and Junhao Liu and Devin Koestler},
   year         = {2017},
   note         = {R package version 1.0},
-  url          = {https://CRAN.R-project.org/package=pwr2}
+  url          = {https://CRAN.R-project.org/package=pwr2},
+  doi          = {10.32614/CRAN.package.pwr2}
 }
 
 @Manual{bulus2025pwrss,
-  title        = {pwrss: Statistical Power and Sample Size Calculation Tools},
+  title        = {{pwrss}: Statistical Power and Sample Size Calculation Tools},
   author       = {Metin Bulus and Sebastian Jentschke},
   year         = {2025},
   note         = {R package version 1.0.0},
-  url          = {https://CRAN.R-project.org/package=pwrss}
+  url          = {https://CRAN.R-project.org/package=pwrss},
+  doi          = {10.32614/CRAN.package.pwrss}
 }
 
 @book{zhang2018webpower,
@@ -81,7 +84,8 @@
   year      = {1988},
   edition   = {2nd},
   publisher = {Lawrence Erlbaum Associates},
-  address   = {Hillsdale, NJ}
+  address   = {Hillsdale, NJ},
+  doi       = {10.1016/C2013-0-10517-X}
 }
 
 @Manual{R,
@@ -90,5 +94,6 @@
   organization = {R Foundation for Statistical Computing},
   address      = {Vienna, Austria},
   year         = {2025},
-  url          = {https://www.R-project.org/}
+  url          = {https://www.R-project.org/},
+  doi          = {10.32614/R.manuals}
 }
