+++ llms.txt
@@ -325,3 +325,9 @@ analyses. Behavior Research Methods, *41*(4), 1149–1160.
 - [`pwrttest()`](https://mutopsy.github.io/pwranova/reference/pwrttest.md)
   : Power Analysis for One-/Two-Sample and Paired t Tests
 
+# Articles
+
+### All vignettes
+
+- [pwranovaパッケージの説明
+  (日本語版)](https://mutopsy.github.io/pwranova/articles/README.ja.md):
