### 2026-07-25 854695e
Preserve capitalization of WebPower and R
