+++ paper/paper.bib
@@ -37,10 +37,10 @@
 }
 
 @book{zhang2018webpower,
-  title   = {Practical Statistical Power Analysis Using Webpower and R},
-  author  = {Zhiyong Zhang and Ke-Hai Yuan},
-  year    = {2018},
-  doi     = {10.35566/power},
+  title     = {Practical Statistical Power Analysis Using {WebPower} and {R}},
+  author    = {Zhiyong Zhang and Ke-Hai Yuan},
+  year      = {2018},
+  doi       = {10.35566/power},
   publisher = {ISDSA Press},
   address   = {Granger, IN}
 }
