### 2026-07-26 4eb2f4a
Added Japanese documentation to the package website
