+++ articles/index.md
@@ -1,6 +1,6 @@
 # Articles
 
-### All vignettes
+### For Japanese users
 
 - [pwranovaパッケージの説明
   (日本語版)](https://mutopsy.github.io/pwranova/articles/README.ja.md):
