+++ authors.md
@@ -10,14 +10,17 @@
 Source:
 [`inst/CITATION`](https://github.com/mutopsy/pwranova/blob/main/inst/CITATION)
 
-Muto, H. (2025). pwranova: An R package for power analysis of flexible
-ANOVA designs and related tests. Jxiv.
-https://doi.org/10.51094/jxiv.1555
+Muto, H. (2026). pwranova: An R package for power analysis of flexible
+ANOVA designs and related tests. Journal of Open Source Software,
+11(123), 10123. https://doi.org/10.21105/joss.10123
 
     @Article{,
       title = {pwranova: An R package for power analysis of flexible ANOVA designs and related tests},
       author = {Hiroyuki Muto},
-      journal = {Jxiv},
-      year = {2025},
-      doi = {10.51094/jxiv.1555},
+      journal = {Journal of Open Source Software},
+      volume = {11},
+      number = {123},
+      pages = {10123},
+      year = {2026},
+      doi = {10.21105/joss.10123},
     }
