### 2026-07-26 cc787ff
Add link to Japanese Documentation
